Creado un controlador para gestionar la tabla. En dicho controlador se crea la propiedad y el método init. Se soluciona el tipo devuelto para los getters del modelo WineryModel.

+++ Baccus/FDCWineryModel.m
@@ -20,16 +20,16 @@
 
 #pragma mark - Properties
 
--(int) redWineCount {
+-(NSUInteger) redWineCount {
     return [self.redWines count];
 }
 
--(int) whiteWineCount
+-(NSUInteger) whiteWineCount
 {
     return [self.whiteWines count];
 }
 
--(int) otherWineCount
+-(NSUInteger) otherWineCount
 {
     return [self.otherWines count];
 }
